pyproject.toml
[project]
name = "dash"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "flask>=3.1.2",
    "flask-migrate>=4.1.0",
    "flask-security-too>=5.7.1",
    "flask-sqlalchemy>=3.1.1",
    "flask-whooshee>=0.9.1",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "yfinance>=1.0",
]
